Add function resetInstance
- Added function resetInstance to set the instance variable of the Database class to null.
Did this in case instance does not have a valid connection to a database. This can happen
if a user tries to login to the app with invalid credentials. If this happens the instance
variable will not be null but will not have a valid connection preventing the user from
logging in again with valid credentials.