Add function compareTransactionDates
- Added function compareTransactionDates which compares two transactions by date to check which transaction
occurred first. The purpose of this function is to be used in order to sort the transaction array list. This
is needed in order for the line chart to function properly. As the line chart will display the rolling
income/expense totals over a given time period which must be ordered.

- Added functionality which sorts the transactions by date