Modify function generatePieChart
- Modified the generatePieChart function to generate the pie chart using the category totals that
are stored in the hashmap. The function no longer iterates through the transactions. Did this so that
when we have to generate the pie charts and line charts each function is not iterating through the same
data as its a waste of time. So I moved alot of the logic outside of the function so we will only have to
iterate through the transactions data one time for all charts to be generated.